Run more test_concurrent_futures tests
In the process, call `quilt push -a`, which pushes all patches,
which is what I meant to do originally.

This imports a fix for https://bugs.python.org/issue40692
that hasn't been formally reviewed yet, but seems reasonable
to me.

It mostly only changes the test suite, but it does also fix
the `compileall` module in a tiny but useful way.